+++ web/app/components/style-guide/data/date-picker.jsx
@@ -41,8 +41,58 @@ var Example = React.createClass({
           </FlexCell>
           <FlexCell>
             <DatePicker
+              label="My Label"
               selectedDay={this.state.selectedDay2}
               onClickDate={this.onClickDate2} />
+          </FlexCell>
+        </FlexRow>
+        <div>
+          just some content after the date picker for visual testing purposes
+        </div>
+      </span>
+    );
+  }
+});
+
+var ExampleWeekMode = React.createClass({
+  getInitialState: function() {
+    return {
+      selectedStartOfWeek1: '01/18/2015',
+      selectedStartOfWeek2: '',
+    };
+  },
+
+  onClickDate1: function(value) {
+    this.setState({
+      selectedStartOfWeek1: value
+    });
+  },
+
+  onClickDate2: function(value) {
+    this.setState({
+      selectedStartOfWeek2: value
+    });
+  },
+
+  render: function() {
+    return (
+      <span>
+        <FlexRow>
+          <FlexCell>
+            <div>selected start of week: {this.state.selectedStartOfWeek1}</div>
+            <DatePicker
+              selectionUnit="week"
+              onClickDate={this.onClickDate1}
+              selectedDay={this.state.selectedStartOfWeek1}
+            />
+          </FlexCell>
+          <FlexCell>
+            <div>selected start of week: {this.state.selectedStartOfWeek2}</div>
+            <DatePicker
+              selectionUnit="week"
+              onClickDate={this.onClickDate2}
+              selectedDay={this.state.selectedStartOfWeek2}
+            />
           </FlexCell>
         </FlexRow>
         <div>
@@ -65,11 +115,21 @@ module.exports = {
   examples: [{
     description: (
       <p>
-        Standard badge.
+        Standard.
       </p>
     ),
     example: (
       <Example />
+    ),
+    exampleString: '<Badge>Standard</Badge>'
+  }, {
+    description: (
+      <p>
+        Week mode.
+      </p>
+    ),
+    example: (
+      <ExampleWeekMode />
     ),
     exampleString: '<Badge>Standard</Badge>'
   }]